refactor(a2a): split adapter by protocol boundary
